# docs/swagger.yaml
components:
  schemas:
...
    handlers.updateUserDTO:
      properties:
        displayname:
          type: string
      type: object
    models.Error:
      properties:
        error:
          type: string
        message:
          type: string
      type: object
    models.Segment:
      properties:
        end:
          type: number
        score:
          type: number
        start:
          type: number
        text:
          type: string
      type: object
    models.Timing:
      properties:
        segments:
          items:
            $ref: '#/components/schemas/models.Segment'
          type: array
          uniqueItems: false
        text:
          type: string
      type: object
...
    sqlc.UserDeleteUserRow:
      properties:
        displayname:
          type: string
        username:
          type: string
      type: object
    sqlc.UserSelectUsersRow:
      properties:
        displayname:
          type: string
        username:
          type: string
      type: object
    sqlc.UserUpdateUserRow:
      properties:
        displayname:
          type: string
        username:
          type: string
      type: object
...
info:
  title: tilawah-hub
  version: 0.0.1
openapi: 3.1.0
paths:
...
  /recitation-timings/{slug}/{verse_key}:
    delete:
      parameters:
      - description: CSRF Token
        in: header
        name: X-CSRF-TOKEN
        required: true
        schema:
          type: string
      - description: Slug
        in: path
        name: slug
        required: true
        schema:
          type: string
      - description: Verse Key
        in: path
        name: verse_key
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.Timing'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.Error'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.Error'
          description: Unauthorized
      tags:
      - RecitationTiming
    post:
      parameters:
      - description: CSRF Token
        in: header
        name: X-CSRF-TOKEN
        required: true
        schema:
          type: string
      - description: Slug
        in: path
        name: slug
        required: true
        schema:
          type: string
      - description: Verse Key
        in: path
        name: verse_key
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/models.Timing'
        description: Update Recitation Timing
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.Timing'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.Error'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.Error'
          description: Unauthorized
      tags:
      - RecitationTiming
...
  /user:
    delete:
      parameters:
      - description: CSRF Token
        in: header
        name: X-CSRF-TOKEN
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/sqlc.UserDeleteUserRow'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.Error'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.Error'
          description: Unauthorized
      tags:
      - User
    put:
      parameters:
      - description: CSRF Token
        in: header
        name: X-CSRF-TOKEN
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/handlers.updateUserDTO'
        description: Update Recitation
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/sqlc.UserUpdateUserRow'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.Error'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.Error'
          description: Unauthorized
      tags:
      - User
  /users:
    get:
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/sqlc.UserSelectUsersRow'
                type: array
          description: OK
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.Error'
          description: Internal Server Error
      tags:
      - User
  /users/{username}:
    get:
      parameters:
      - description: Username
        in: path
        name: username
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/sqlc.UserSelectUsersRow'
                type: array
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/models.Error'
          description: Bad Request
      tags:
      - User